move network state to global bus
