+++ packages/app/src/contexts/Network/index.tsx
@@ -5,9 +5,9 @@ import { createSafeContext } from '@w3ux/hooks'
 import { varToUrlHash } from '@w3ux/utils'
 import { NetworkList } from 'consts/networks'
 import { Apis } from 'controllers/Apis'
-import { getInitialNetwork } from 'global-bus/util'
+import { getNetwork, network$, setNetwork } from 'global-bus'
 import type { ReactNode } from 'react'
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
 import type { NetworkId } from 'types'
 import type { NetworkContextInterface, NetworkState } from './types'
 
@@ -15,6 +15,12 @@ export const [NetworkContext, useNetwork] =
   createSafeContext<NetworkContextInterface>()
 
 export const NetworkProvider = ({ children }: { children: ReactNode }) => {
+  // Store the active network in state
+  const [network, setNetworkState] = useState<NetworkState>({
+    name: getNetwork(),
+    meta: NetworkList[getNetwork()],
+  })
+
   // handle network switching
   const switchNetwork = async (name: NetworkId): Promise<void> => {
     // Disconnect from current APIs before switching network
@@ -22,23 +28,22 @@ export const NetworkProvider = ({ children }: { children: ReactNode }) => {
       await Apis.destroy(network.name),
       await Apis.destroy(`people-${network.name}`),
     ])
-
-    setNetwork({
-      name,
-      meta: NetworkList[name],
-    })
-
-    // update url `n` if needed
+    setNetwork(name)
     varToUrlHash('n', name, false)
   }
 
-  // Store the initial active network
-  const initialNetwork = getInitialNetwork()
-
-  const [network, setNetwork] = useState<NetworkState>({
-    name: initialNetwork,
-    meta: NetworkList[initialNetwork],
-  })
+  // Subscribe to global bus network changes
+  useEffect(() => {
+    const sub = network$.subscribe((n) => {
+      setNetworkState({
+        name: n,
+        meta: NetworkList[n],
+      })
+    })
+    return () => {
+      sub.unsubscribe()
+    }
+  }, [])
 
   return (
     <NetworkContext.Provider
