+++ packages/global-bus/src/network/index.ts
@@ -1,0 +1,12 @@
+// Discovered extensions along with their status
+
+import type { NetworkId } from 'types'
+import { _network } from './private'
+
+export const network$ = _network.asObservable()
+
+export const getNetwork = () => _network.getValue()
+
+export const setNetwork = (network: NetworkId) => {
+  _network.next(network)
+}
